+++ packages/vue/framework7-vue.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 Vue 3.6.0
+ * Framework7 Vue 3.6.1
  * Build full featured iOS & Android apps using Framework7 & Vue
  * http://framework7.io/vue/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: December 7, 2018
+ * Released on: December 10, 2018
  */
 
 (function (global, factory) {
@@ -10161,17 +10161,19 @@
       var className = props.className;
       var routable = props.routable;
       var classes = Utils.classNames(className, Mixins.colorClasses(props));
-      var tabsClasses = Utils.classNames({
-        'tabs': true,
+      var wrapClasses = Utils.classNames({
         'tabs-animated-wrap': animated,
-        'tabs-swipeable-wrap': swipeable,
+        'tabs-swipeable-wrap': swipeable
+      });
+      var tabsClasses = Utils.classNames({
+        tabs: true,
         'tabs-routable': routable
       });
 
       if (animated || swipeable) {
         return _h('div', {
           style: style,
-          class: classes,
+          class: Utils.classNames(wrapClasses, classes),
           attrs: {
             id: id
           }
@@ -10299,7 +10301,7 @@
       url: String,
       main: Boolean,
       stackPages: Boolean,
-      xhrCache: String,
+      xhrCache: Boolean,
       xhrCacheIgnore: Array,
       xhrCacheIgnoreGetParameters: Boolean,
       xhrCacheDuration: Number,
@@ -10768,7 +10770,7 @@
   };
 
   /**
-   * Framework7 Vue 3.6.0
+   * Framework7 Vue 3.6.1
    * Build full featured iOS & Android apps using Framework7 & Vue
    * http://framework7.io/vue/
    *
@@ -10776,7 +10778,7 @@
    *
    * Released under the MIT License
    *
-   * Released on: December 7, 2018
+   * Released on: December 10, 2018
    */
 
   var Plugin = {