+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Change Log
 
+# [v3.6.1](https://github.com/framework7io/framework7/compare/v3.6.0...v3.6.1) - December 10, 2018
+  * Phenome (React / Vue)
+    * Tabs - fixed issue with broken Animated/Swipeable tabs
+
 # [v3.6.0](https://github.com/framework7io/framework7/compare/v3.5.2...v3.6.0) - December 7, 2018
   * Core
     * Router